Change PID example from braking car to heater heating a room
The heater example applies better to most applications of PID, with an error that can go above and below the set point. Also the heater affects temperature, to get to a set point of temperature. The breaking example has breaking, which affects acceleration, which affects speed, which affects position, but has a set point of position.

* Add graph of heater

* Add heater graph

* Remove file extension to commit lowercase

Just changing the file extension to lowercase (.png from .PNG) means it
won't get detected for commiting. So I'm removing the extension, then
will add it back, lowercase, in the next commit.

* Add lowercase png extension to heater Temp Graph

see previous commit message for details

* Add lowercase png extension to heater Temp Graph

see previous commit message for details